improve-cli-build-commands: refactor manager.go into smaller files
Split manager.go (1444 lines) into focused modules:
- build.go: Build-related functions (RunBuild, buildLocally, buildWithPodman)
- install.go: Install functions for all platforms
- deploy.go: Deployment to remote robots
- ssh_utils.go: SSH connection and file transfer utilities
- provision.go: Tailscale key provisioning
- remote.go: Remote operations (logs, sync-code, remote-build)
- config.go: Configuration loading and utility functions

All 14 tests passing
